requirements.txt
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
idna==2.8
psycopg2-binary==2.8.4
pycparser==2.19
pyOpenSSL==19.1.0
PySocks==1.7.1
requests==2.22.0
six==1.13.0
urllib3==1.25.8
win-inet-pton==1.1.0
wincertstore==0.2
